+++ src/models/card.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MemnixCard {
-    pub id: i8,
+    pub id: u32,
     pub question: String,
     pub answer: String,
     //pub deck_id: i64,
